python/requirements.txt
argparse>=1.4.0
py-color-log>=1.0.4
chardet>=5.1.0
pyDes>=2.0.1
pycryptodome>=3.17